requirements.txt
flask
python-dotenv
groq
requests>=2.31.0
flask-cors